- change pedersen to ecc - move edwards_d to top level
