+++ src/constraint_system/mod.rs
@@ -10,12 +10,12 @@ pub(crate) mod variable;
 pub mod arithmetic;
 /// Boolean gate
 pub mod boolean;
+/// Scalar multiplication gate
+pub mod ecc;
 #[cfg(test)]
 pub(crate) mod helper;
 /// XOR and AND gates
 pub mod logic;
-/// Scalar multiplication gate
-pub mod pedersen;
 /// Range gate
 pub mod range;
