+++ src/proof_system/widget/ecc/mod.rs
@@ -5,22 +5,16 @@ pub use proverkey::ProverKey;
 pub use verifierkey::VerifierKey;
 
 // Note: The ECC gadget does not check that the initial point is on the curve for two reasons:
-//  - We constrain the accumulator to start from the identity point, which the verifier knows is on the curve
+// - We constrain the accumulator to start from the identity point, which the verifier knows is on the curve
 // - We are adding multiples of the generator to the accumulator which the verifier also knows is on the curve and is prime order
 // - We do allow arbitrary Scalar multiplication, and possibly XXX: may add constraints to ensure the generator is correct (prime order)
 
 // Bits are accumulated in base2. So we use d(Xw) - 2d(X) to extract the base2 bit
+use crate::edwards_d;
 use dusk_bls12_381::Scalar;
-use jubjub::Fq;
 fn extract_bit(curr_acc: &Scalar, next_acc: &Scalar) -> Scalar {
     // Next - 2 * current
     next_acc - (curr_acc + curr_acc)
-}
-
-fn edwards_d() -> Fq {
-    let num = Fq::from(10240);
-    let den = Fq::from(10241);
-    -(num * den.invert().unwrap())
 }
 
 // Ensures that the bit is either +1, -1 or 0
